requirements.txt
# Core dependencies
chainlit>=2.0.0
litellm>=1.30.0
openai>=1.20.0
pydantic==2.10.1  # Specific version due to compatibility issues with chainlit: https://github.com/Chainlit/chainlit/issues/1556
python-dotenv>=1.0.0
supabase>=2.15.0

# Semantic routing
semantic-router[fastembed]>=0.0.54

# Image processing
pillow>=10.0.0

# AI models
google-generativeai>=0.3.0

# Web search
tavily-python

# Documentation dependencies (optional)
# mkdocs-llmstxt>=0.2.0

# Development dependencies (use pip install -e ".[dev]" to install these)
# pytest>=8.3.5
# pytest-asyncio>=0.26.0
# twine>=6.1.0
# black>=23.7.0
# isort>=5.12.0
# flake8>=6.1.0
# ruff>=0.11.4
